move osmo message related utils to osmosis-utils package
