+++ packages/osmosis-utils/Cargo.toml
@@ -17,3 +17,4 @@ cosmwasm-schema         = { workspace = true }
 osmosis-test-tube       = { workspace = true, optional = true }
 cosmwasm-std-polytone   = { package = "cosmwasm-std", version = "1.5.7" }
 valence-account-utils   = { workspace = true }
+osmosis-std             = { workspace = true }
